requirements.txt
streamlit
faiss-cpu
sentence-transformers
numpy
openai>=1.51.0
